Remoção do comando Snipe

+++ src/main/java/com/kuuhaku/events/guild/GuildEvents.java
@@ -285,17 +285,23 @@ public class GuildEvents extends ListenerAdapter {
 	}
 
 	private void countSpam(Member member, MessageChannel channel, Guild guild, List<Message> h) {
-		if (h.size() >= GuildDAO.getGuildById(guild.getId()).getNoSpamAmount()) {
+		if (h.size() >= GuildDAO.getGuildById(guild.getId()).getNoSpamAmount() && Helper.hasRoleHigherThan(guild.getSelfMember(), member)) {
 			h.forEach(m -> channel.deleteMessageById(m.getId()).complete());
 			channel.sendMessage(":warning: | Opa, sem spam meu amigo!").queue(
 					msg -> msg.delete().queueAfter(20, TimeUnit.SECONDS)
 			);
 			try {
 				Role r = guild.getRoleById(GuildDAO.getGuildById(guild.getId()).getCargoWarn());
-				if (r != null) guild.addRoleToMember(member, r)
-						.delay(GuildDAO.getGuildById(guild.getId()).getWarnTime(), TimeUnit.MINUTES)
-						.flatMap(s -> guild.removeRoleFromMember(member, r))
-						.queue();
+				if (r != null) {
+					guild.addRoleToMember(member, r)
+							.delay(GuildDAO.getGuildById(guild.getId()).getWarnTime(), TimeUnit.MINUTES)
+							.flatMap(s -> guild.removeRoleFromMember(member, r))
+							.queue();
+					MutedMember mm = Helper.getOr(com.kuuhaku.controller.postgresql.MemberDAO.getMutedMemberById(member.getId()), new MutedMember(member.getId(), guild.getId()));
+					mm.mute(GuildDAO.getGuildById(guild.getId()).getWarnTime());
+
+					com.kuuhaku.controller.postgresql.MemberDAO.saveMutedMember(mm);
+				}
 			} catch (Exception ignore) {
 			}
 		}
